// src/types/generated/Note.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TextFormat } from "./TextFormat";

export type Note = { id: string, projectId: string | null, title: string, body: string, bodyFormat: TextFormat, sortOrder: number, createdAt: string, updatedAt: string, deletedAt: string | null, };